+++ src/Controller/PagesController.php
@@ -83,9 +83,7 @@ class PagesController extends AppController
 
         $page = $this->Pages->find('active')
             ->select(['id', 'title', 'subtitle', 'slug', 'text', 'active', 'created', 'modified'])
-            ->contain(['Categories' => function ($q) {
-                return $q->select(['title', 'slug']);
-            }])
+            ->contain(['Categories' => ['fields' => ['title', 'slug']]])
             ->where([sprintf('%s.slug', $this->Pages->getAlias()) => $slug])
             ->cache(sprintf('view_%s', md5($slug)), $this->Pages->cache)
             ->firstOrFail();
@@ -103,9 +101,7 @@ class PagesController extends AppController
     {
         $page = $this->Pages->find('pending')
             ->select(['id', 'title', 'subtitle', 'slug', 'text', 'active', 'created', 'modified'])
-            ->contain(['Categories' => function ($q) {
-                return $q->select(['title', 'slug']);
-            }])
+            ->contain(['Categories' => ['fields' => ['title', 'slug']]])
             ->where([sprintf('%s.slug', $this->Pages->getAlias()) => $slug])
             ->firstOrFail();