+++ src/Controller/PhotosAlbumsController.php
@@ -39,7 +39,7 @@ class PhotosAlbumsController extends AppController
     {
         $albums = $this->PhotosAlbums->find('active')
             ->select(['id', 'title', 'slug', 'photo_count'])
-            ->contain(['Photos' => function ($q) {
+            ->contain([$this->PhotosAlbums->Photos->getAlias() => function ($q) {
                 return $q->find('active')
                     ->select(['album_id', 'filename'])
                     ->order('rand()');
@@ -110,30 +110,5 @@ class PhotosAlbumsController extends AppController
         }
 
         $this->set(compact('album', 'photos'));
-    }
-
-    /**
-     * Preview for albums.
-     * It uses the `view` template.
-     * @param string $slug Album slug
-     * @return \Cake\Network\Response
-     */
-    public function preview($slug = null)
-    {
-        $album = $this->PhotosAlbums->find()
-            ->select(['id', 'slug', 'title', 'active'])
-            ->contain(['Photos' => function ($q) {
-                return $q->select(['id', 'album_id', 'filename', 'description'])
-                    ->order([
-                        sprintf('%s.created', $this->PhotosAlbums->Photos->getAlias()) => 'DESC',
-                        sprintf('%s.id', $this->PhotosAlbums->Photos->getAlias()) => 'DESC',
-                    ]);
-            }])
-            ->where(compact('slug'))
-            ->firstOrFail();
-
-        $this->set(compact('album'));
-
-        $this->render('view');
     }
 }