+++ tests/TestCase/Controller/PhotosControllerTest.php
@@ -1,0 +1,129 @@
+<?php
+/**
+ * This file is part of MeCms.
+ *
+ * MeCms is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * MeCms is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with MeCms.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
+ * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
+ * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
+ * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ */
+namespace MeCms\Test\TestCase\Controller;
+
+use Cake\Cache\Cache;
+use Cake\ORM\TableRegistry;
+use Cake\TestSuite\IntegrationTestCase;
+
+/**
+ * PhotosControllerTest class
+ */
+class PhotosControllerTest extends IntegrationTestCase
+{
+    /**
+     * @var \MeCms\Model\Table\PhotosTable
+     */
+    protected $Photos;
+
+    /**
+     * Fixtures
+     * @var array
+     */
+    public $fixtures = [
+        'plugin.me_cms.photos',
+        'plugin.me_cms.photos_albums',
+    ];
+
+    /**
+     * Setup the test case, backup the static object values so they can be
+     * restored. Specifically backs up the contents of Configure and paths in
+     *  App if they have not already been backed up
+     * @return void
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->Photos = TableRegistry::get('MeCms.Photos');
+
+        Cache::clear(false, $this->Photos->cache);
+    }
+
+    /**
+     * Teardown any static object changes and restore them
+     * @return void
+     */
+    public function tearDown()
+    {
+        parent::tearDown();
+
+        unset($this->Photos);
+    }
+
+    /**
+     * Adds additional event spies to the controller/view event manager
+     * @param \Cake\Event\Event $event A dispatcher event
+     * @param \Cake\Controller\Controller|null $controller Controller instance
+     * @return void
+     */
+    public function controllerSpy($event, $controller = null)
+    {
+        $controller->viewBuilder()->setLayout(false);
+
+        parent::controllerSpy($event, $controller);
+    }
+
+    /**
+     * Tests for `view()` method
+     * @test
+     */
+    public function testView()
+    {
+        $id = 1;
+        $url = ['_name' => 'photo', 'test-album', $id];
+
+        $this->get($url);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Photos/view.ctp');
+
+        $photoFromView = $this->viewVariable('photo');
+        $this->assertInstanceof('MeCms\Model\Entity\Photo', $photoFromView);
+
+        $cache = Cache::read(sprintf('view_%s', md5($id)), $this->Photos->cache);
+        $this->assertEquals($photoFromView, $cache->first());
+
+        //Backward compatibility for URLs like `/photo/11`
+        $this->get('/photo/' . $id);
+        $this->assertRedirect($url);
+    }
+
+    /**
+     * Tests for `preview()` method
+     * @test
+     */
+    public function testPreview()
+    {
+        $id = $this->Photos->find('pending')->extract('id')->first();
+        $url = ['_name' => 'photosPreview', $id];
+
+        $this->get($url);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Photos/view.ctp');
+
+        $photoFromView = $this->viewVariable('photo');
+        $this->assertInstanceof('MeCms\Model\Entity\Photo', $photoFromView);
+    }
+}